Provisioning: Share one Grafana server per package in repository integration tests
Apply the shared Grafana server strategy (sync.Once + TestMain) to the
repository test package, following the same pattern established for the
connection package in #120906.

Server startups drop from 25 → 4 (1 shared + 3 with custom options).

- Rewrite helpers_test.go with common.NewSharedEnv() + sharedHelper + TestMain
- Replace common.RunGrafana(t) with sharedHelper(t) in 21 tests
- Remove testutil.SkipIntegrationTestInShortMode (handled by GetHelper)
- Move TestIntegrationRepositoryController_EnterpriseWiring to enterprise/
- Keep 3 tests with custom GrafanaOptions on their own servers

Made-with: Cursor
