+++ pkg/tests/apis/provisioning/enterprise/enterprise_test.go
@@ -571,3 +571,187 @@ func TestIntegrationConnectionController_EnterpriseWiring(t *testing.T) {
 		}
 	})
 }
+
+func TestIntegrationRepositoryController_EnterpriseWiring(t *testing.T) {
+	helper := sharedHelper(t)
+	ctx := context.Background()
+
+	t.Run("GitLab repository can be created and reconciled", func(t *testing.T) {
+		token := base64.StdEncoding.EncodeToString([]byte("test-gitlab-token"))
+
+		repository := &unstructured.Unstructured{Object: map[string]any{
+			"apiVersion": "provisioning.grafana.app/v0alpha1",
+			"kind":       "Repository",
+			"metadata": map[string]any{
+				"name":      "test-gitlab-repo",
+				"namespace": "default",
+			},
+			"spec": map[string]any{
+				"title": "Test GitLab Repository",
+				"type":  string(provisioning.GitLabRepositoryType),
+				"gitlab": map[string]any{
+					"url":    "https://gitlab.com/test/repo.git",
+					"branch": "main",
+					"path":   "dashboards",
+				},
+			},
+			"secure": map[string]any{
+				"token": map[string]any{
+					"create": token,
+				},
+			},
+		}}
+
+		created, err := helper.Repositories.Resource.Create(ctx, repository, metav1.CreateOptions{})
+		require.NoError(t, err, "failed to create GitLab repository")
+		require.NotNil(t, created)
+
+		repoName := created.GetName()
+		require.NotEmpty(t, repoName, "repository name should not be empty")
+
+		t.Cleanup(func() {
+			_ = helper.Repositories.Resource.Delete(ctx, repoName, metav1.DeleteOptions{})
+		})
+
+		output, err := helper.Repositories.Resource.Get(ctx, repoName, metav1.GetOptions{})
+		require.NoError(t, err, "failed to read back GitLab repository")
+		assert.Equal(t, repoName, output.GetName(), "name should be equal")
+
+		spec := output.Object["spec"].(map[string]any)
+		assert.Equal(t, string(provisioning.GitLabRepositoryType), spec["type"], "type should be gitlab")
+
+		restConfig := helper.Org1.Admin.NewRestConfig()
+		provClient, err := clientset.NewForConfig(restConfig)
+		require.NoError(t, err, "failed to create provisioning client")
+		repoClient := provClient.ProvisioningV0alpha1().Repositories("default")
+
+		require.Eventually(t, func() bool {
+			updated, err := repoClient.Get(ctx, repoName, metav1.GetOptions{})
+			if err != nil {
+				return false
+			}
+			return updated.Status.ObservedGeneration == updated.Generation &&
+				updated.Status.Health.Checked > 0
+		}, 15*time.Second, 500*time.Millisecond, "repository should be reconciled by controller")
+
+		reconciled, err := repoClient.Get(ctx, repoName, metav1.GetOptions{})
+		require.NoError(t, err)
+
+		assert.Equal(t, reconciled.Generation, reconciled.Status.ObservedGeneration,
+			"controller should have reconciled the repository")
+		assert.Greater(t, reconciled.Status.Health.Checked, int64(0),
+			"health check should have been attempted")
+
+		readyCondition := meta.FindStatusCondition(reconciled.Status.Conditions, provisioning.ConditionTypeReady)
+		assert.NotNil(t, readyCondition, "should have ready condition")
+
+		t.Logf("GitLab repository reconciled successfully. Health: %v, ObservedGen: %d, Checked: %d",
+			reconciled.Status.Health.Healthy, reconciled.Status.ObservedGeneration, reconciled.Status.Health.Checked)
+	})
+
+	t.Run("Bitbucket repository can be created and reconciled", func(t *testing.T) {
+		token := base64.StdEncoding.EncodeToString([]byte("test-bitbucket-token"))
+
+		repository := &unstructured.Unstructured{Object: map[string]any{
+			"apiVersion": "provisioning.grafana.app/v0alpha1",
+			"kind":       "Repository",
+			"metadata": map[string]any{
+				"name":      "test-bitbucket-repo",
+				"namespace": "default",
+			},
+			"spec": map[string]any{
+				"title": "Test Bitbucket Repository",
+				"type":  string(provisioning.BitbucketRepositoryType),
+				"bitbucket": map[string]any{
+					"url":    "https://bitbucket.org/workspace/repo.git",
+					"branch": "main",
+					"path":   "dashboards",
+				},
+			},
+			"secure": map[string]any{
+				"token": map[string]any{
+					"create": token,
+				},
+			},
+		}}
+
+		created, err := helper.Repositories.Resource.Create(ctx, repository, metav1.CreateOptions{})
+		require.NoError(t, err, "failed to create Bitbucket repository")
+		require.NotNil(t, created)
+
+		repoName := created.GetName()
+		require.NotEmpty(t, repoName, "repository name should not be empty")
+
+		t.Cleanup(func() {
+			_ = helper.Repositories.Resource.Delete(ctx, repoName, metav1.DeleteOptions{})
+		})
+
+		output, err := helper.Repositories.Resource.Get(ctx, repoName, metav1.GetOptions{})
+		require.NoError(t, err, "failed to read back Bitbucket repository")
+		assert.Equal(t, repoName, output.GetName(), "name should be equal")
+
+		spec := output.Object["spec"].(map[string]any)
+		assert.Equal(t, string(provisioning.BitbucketRepositoryType), spec["type"], "type should be bitbucket")
+
+		restConfig := helper.Org1.Admin.NewRestConfig()
+		provClient, err := clientset.NewForConfig(restConfig)
+		require.NoError(t, err, "failed to create provisioning client")
+		repoClient := provClient.ProvisioningV0alpha1().Repositories("default")
+
+		require.Eventually(t, func() bool {
+			updated, err := repoClient.Get(ctx, repoName, metav1.GetOptions{})
+			if err != nil {
+				return false
+			}
+			return updated.Status.ObservedGeneration == updated.Generation &&
+				updated.Status.Health.Checked > 0
+		}, 15*time.Second, 500*time.Millisecond, "repository should be reconciled by controller")
+
+		reconciled, err := repoClient.Get(ctx, repoName, metav1.GetOptions{})
+		require.NoError(t, err)
+
+		assert.Equal(t, reconciled.Generation, reconciled.Status.ObservedGeneration,
+			"controller should have reconciled the repository")
+		assert.Greater(t, reconciled.Status.Health.Checked, int64(0),
+			"health check should have been attempted")
+
+		readyCondition := meta.FindStatusCondition(reconciled.Status.Conditions, provisioning.ConditionTypeReady)
+		assert.NotNil(t, readyCondition, "should have ready condition")
+
+		t.Logf("Bitbucket repository reconciled successfully. Health: %v, ObservedGen: %d, Checked: %d",
+			reconciled.Status.Health.Healthy, reconciled.Status.ObservedGeneration, reconciled.Status.Health.Checked)
+	})
+
+	t.Run("All repository types are supported", func(t *testing.T) {
+		supportedTypes := []provisioning.RepositoryType{
+			provisioning.GitHubRepositoryType,
+			provisioning.GitLabRepositoryType,
+			provisioning.BitbucketRepositoryType,
+			provisioning.GitRepositoryType,
+			provisioning.LocalRepositoryType,
+		}
+
+		for _, repoType := range supportedTypes {
+			t.Run(string(repoType), func(t *testing.T) {
+				repo := &unstructured.Unstructured{Object: map[string]any{
+					"apiVersion": "provisioning.grafana.app/v0alpha1",
+					"kind":       "Repository",
+					"metadata": map[string]any{
+						"generateName": "test-",
+						"namespace":    "default",
+					},
+					"spec": map[string]any{
+						"title": "Test Repository",
+						"type":  string(repoType),
+					},
+				}}
+
+				_, err := helper.Repositories.Resource.Create(ctx, repo, metav1.CreateOptions{})
+				if err != nil {
+					assert.NotContains(t, err.Error(), "is not supported",
+						"type %s should be supported by factory", repoType)
+				}
+			})
+		}
+	})
+}
